Correios Rev0
First release

This code receives the tracking code, checks the status, and then checks again after a calibratable amount of time. When a change is noticed, the program ends and the final status is shown. Code still needs optimization, a way of saving data into a dataset (maybe sourcing the tracking codes from one dataset also), and learn something from this dataset.